refactor: move static string into templates/ folder
