+++ crates/tuono/src/source_builder.rs
@@ -13,70 +13,10 @@ use crate::mode::Mode;
 use crate::route::AxumInfo;
 use crate::route::Route;
 
-const FALLBACK_HTML: &str = r#"<!doctype html>
-<html>
-<body>
-  <script>
-    window['__TUONO_SERVER_PAYLOAD__'] = [SERVER_PAYLOAD]
-  </script>
-  <script type="module" async>
-    import RefreshRuntime from '[BASE_URL]/vite-server/@react-refresh'
-    RefreshRuntime.injectIntoGlobalHook(window)
-    window.$RefreshReg$ = () => { }
-    window.$RefreshSig$ = () => (type) => type
-    window.__vite_plugin_react_preamble_installed__ = true
-  </script>
-  <script type="module" async src="[BASE_URL]/vite-server/@vite/client"></script>
-  <script type="module" async src="[BASE_URL]/vite-server/client-main.tsx"></script>
-</body>
-</html>"#;
-
-const SERVER_ENTRY_DATA: &str = "// File automatically generated by tuono
-// Do not manually update this file
-import { routeTree } from './routeTree.gen'
-import { serverSideRendering } from 'tuono/ssr'
-
-export const renderFn = serverSideRendering(routeTree)
-";
-
-const CLIENT_ENTRY_DATA: &str = "// File automatically generated by tuono
-// Do not manually update this file
-import 'vite/modulepreload-polyfill'
-import { hydrate } from 'tuono/hydration'
-
-// Import the generated route tree
-import { routeTree } from './routeTree.gen'
-
-hydrate(routeTree)
-";
-
-const AXUM_ENTRY_POINT: &str = r##"
-// File automatically generated
-// Do not manually change it
-
-use tuono_lib::{tokio, Mode, Server, axum::Router, tuono_internal_init_v8_platform};
-// AXUM_GET_ROUTE_HANDLER
-
-const MODE: Mode = /*MODE*/;
-
-// MODULE_IMPORTS
-
-//MAIN_FILE_IMPORT//
-
-#[tokio::main]
-async fn main() {
-    tuono_internal_init_v8_platform();
-    println!("\n  ⚡ Tuono v/*VERSION*/");
-
-    //MAIN_FILE_DEFINITION//
-
-    let router = Router::new()
-        // ROUTE_BUILDER
-        //MAIN_FILE_USAGE//;
-
-    Server::init(router, MODE).await.start().await
-}
-"##;
+const FALLBACK_HTML: &str = include_str!("../templates/fallback.html");
+const SERVER_ENTRY_DATA: &str = include_str!("../templates/server.ts");
+const CLIENT_ENTRY_DATA: &str = include_str!("../templates/client.ts");
+const AXUM_ENTRY_POINT: &str = include_str!("../templates/server.rs");
 
 #[cfg(target_os = "windows")]
 const MAIN_FILE_PATH: &str = ".\\.tuono\\main.rs";
