+++ crates/tuono/templates/server.ts
@@ -1,0 +1,7 @@
+// File automatically generated by tuono
+// Do not manually update this file
+import { serverSideRendering } from 'tuono/ssr'
+
+import { routeTree } from './routeTree.gen'
+
+export const renderFn = serverSideRendering(routeTree)
